MTN trust model corrected, sort of.
This still is a bit random as it lives in ~/.monotone/monotonerc and not in a local file, but now the key is verified instead of only the name.

Trust now is modeled such, that only the 3 major devs may commit to head.

In the repository, there are unsigned commits.  This is not a problem, as the head is signed.  And there is one commit which is supposed to warn, it is from "untrusted-test@mail.i2p".  You guess it, it should proof that trust checks are working.

As I found no way to handle differencial syncs from MTN to GIT I will only synchronize both repositories the "natural" (read: manual) way.  There is "mtn git_export", but I did not find a way to control it properly.  And GIT seems not to allow path translation on "git fast import".

There are tools out there (for example "taylor") which allow to import into a different directory, but I do not go down this road today.  So my repository is bugged.  If you create a better one I will be happy to switch ;)